+++ README.md
@@ -1,25 +1,1 @@
 # qa_python
-# Приложение BooksCollector 
-# Оно позволяет установить жанр книг и добавить их в избранное.
-# main.py   Исходный код BooksCollector
-# test_bookscollector.py Автотесты 
-| Добавление книги      | Книга с валидным названием (1-40 символов) успешно добавляется          | `add_new_book`                    |
-|                       | Книга с невалидным названием (0 или >40 символов) не добавляется        | `add_new_book`                    |
-|                       | Невозможность добавления дубликата                                      | `add_new_book`                    |
-| Работа с жанрами      | Установка разрешенного жанра для существующей книги                     | `set_book_genre`, `get_book_genre`|
-|                       | Невозможность установки недопустимого жанра                             | `set_book_genre`                  |
-|                       | Автоматическое отсутствие жанра у новой книги                           | `get_book_genre`                  |
-| Фильтрация            | Получение книг по конкретному жанру                                     | `get_books_with_specific_genre`   |
-|                       | Отсутствие книг с несуществующим жанром в фильтрации                    | `get_books_with_specific_genre`   |
-| Детская подборка      | Исключение книг с возрастным рейтингом (`Ужасы`, `Детективы`)           | `get_books_for_children`          |
-|                       | Включение книг с безопасными жанрами (`Фантастика`, `Мультфильмы` и др.)| `get_books_for_children`          |
-| Избранное             | Успешное добавление/удаление из избранного                             | `add_book_in_favorites`, `delete_book_from_favorites` |
-|                       | Защита от дублирования в избранном                                      | `add_book_in_favorites`           |
-|                       | Невозможность добавления несуществующей книги в избранное               | `add_book_in_favorites`           |
-| Состояние системы     | Корректность начальных значений списков жанров                          | -                                 |
-|                       | Пустая коллекция при инициализации                                      | `get_books_genre`   
-
-# conftest.py   Фикстуры:
-|   collector_with_books   6 книг с разными жанрами
-|   collector_with_favorites  3 книги в избранном
-README.md             Документация
